% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\docType{methods}
\name{createSubMatrix,GTEx-method}
\alias{createSubMatrix,GTEx-method}
\alias{createSubMatrix}
\title{create a tissue-specific expression matrix}
\usage{
\S4method{createSubMatrix}{GTEx}(obj, tissueName)
}
\arguments{
\item{obj}{An instance of the GTEx class}

\item{tissueName}{A character string, like "lung"}
}
\value{
An invisible expression matrix
}
\description{
Given a string which matches all or part of a GTEx tissue name, return the matrix sith geneSymbol
names.
}
