% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\docType{class}
\name{GTEx-class}
\alias{GTEx-class}
\alias{.GTEx}
\alias{GTEx}
\title{GTEx-class}
\usage{
GTEx(quiet = TRUE)
}
\value{
An object of the GTEx class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
