% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\docType{methods}
\name{normalize,GTEx-method}
\alias{normalize,GTEx-method}
\alias{normalize}
\title{normalize}
\usage{
\S4method{normalize}{GTEx}(obj, matrix, strategy)
}
\arguments{
\item{obj}{An instance of the GTEx class}

\item{matrix}{an expression matrix}

\item{strategy}{A character string: "asinh",}
}
\value{
An invisible expression matrix
}
\description{
transform counts using the specified strategy
}
